admin: move the server startup to typescript
The application server's startup is now done from the evolution-backend
package.
